% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\name{geom_density_default}
\alias{geom_density_default}
\title{Default KDE (Density) Function for Rosé Pine Theme}
\usage{
geom_density_default(
  alpha = 0.7,
  fill_color = rose_pine_colors$accent1,
  size = 1.2,
  ...
)
}
\arguments{
\item{alpha}{Transparency level for the fill (default: 0.7).}

\item{fill_color}{Color for the fill (default: Rosé Pine accent1).}

\item{size}{Line size (default: 1.2).}

\item{...}{Additional arguments passed to `geom_density`.}

\item{line_color}{Color for the line (default: Rosé Pine accent5).}
}
\value{
A ggplot2 `geom_density` layer.
}
\description{
A helper function for non-grouped KDE (density) plots with default fill and line color.
}
