% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\name{geom_histogram_default}
\alias{geom_histogram_default}
\title{Default Histogram Function for Rosé Pine Theme}
\usage{
geom_histogram_default(data_length, fill_color = NULL, ...)
}
\arguments{
\item{data_length}{The length of the data to calculate the number of bins.}

\item{fill_color}{The fill color for the histogram bars.}

\item{...}{Additional arguments passed to `geom_histogram`.}
}
\value{
A ggplot2 `geom_histogram` layer.
}
\description{
A helper function for non-grouped histograms with dynamic fill and border colors.
}
