% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\docType{data}
\name{rose_pine_colors}
\alias{rose_pine_colors}
\title{Rosé Pine Colors}
\format{
An object of class \code{list} of length 203.
}
\usage{
rose_pine_colors
}
\description{
A list of Rosé Pine theme colors used for ggplot2 themes and scales.
}
\keyword{datasets}
