% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\name{scale_color_rose_pine}
\alias{scale_color_rose_pine}
\title{Rosé Pine Color Palette}
\usage{
scale_color_rose_pine(...)
}
\arguments{
\item{...}{Additional arguments passed to `scale_color_manual`.}
}
\value{
A scale object.
}
\description{
Default color palette for Rosé Pine theme.
}
