% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\name{scale_fill_rose_pine}
\alias{scale_fill_rose_pine}
\title{Rosé Pine Fill Palette}
\usage{
scale_fill_rose_pine(...)
}
\arguments{
\item{...}{Additional arguments passed to `scale_fill_manual`.}
}
\value{
A scale object.
}
\description{
Default fill palette for Rosé Pine theme.
}
