% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rose_pine.R
\name{theme_rose_pine}
\alias{theme_rose_pine}
\title{Rosé Pine ggplot2 Theme}
\usage{
theme_rose_pine(
  base_size = 12,
  base_family = "Source Code Pro",
  line_color = "#31748f",
  no_margins = FALSE
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family (default: "Source Code Pro").}

\item{line_color}{Color for the lines in geom_line() (default: #31748f for teal).}
}
\value{
A ggplot2 theme object.
}
\description{
A custom ggplot2 theme inspired by the Rosé Pine aesthetic.
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  theme_rose_pine()
}
